+++ src/main/java/com/microsoft/sqlserver/jdbc/Geography.java
@@ -11,6 +11,7 @@ import java.nio.ByteOrder;
 
 /**
  * Geography datatype represents data in a round-earth coordinate system.
+ * This class will stay in this current package for backwards compatibility.
  */
 
 public class Geography extends SQLServerSpatialDatatype {
