+++ src/main/java/com/microsoft/sqlserver/jdbc/SQLServerSpatialDatatype.java
@@ -13,10 +13,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
-import microsoft.sql.Figure;
-import microsoft.sql.Point;
-import microsoft.sql.Segment;
-import microsoft.sql.Shape;
+import com.microsoft.sqlserver.jdbc.spatialdatatypes.Figure;
+import com.microsoft.sqlserver.jdbc.spatialdatatypes.Point;
+import com.microsoft.sqlserver.jdbc.spatialdatatypes.Segment;
+import com.microsoft.sqlserver.jdbc.spatialdatatypes.Shape;
 
 
 /**
